PROJET3/PROJET3/ProjetAzure/functions/requirements.txt
azure-functions
azure-storage-blob
python-multipart
Pillow
